+++ tom_d4rt_ast/lib/src/runtime/bridge/bridged_types.dart
@@ -392,8 +392,9 @@ class BridgedInstance<T extends Object> implements RuntimeValue {
     // would need to be returned as a value.
 
     // 3. If neither method nor getter found, throw an error
-    throw RuntimeD4rtException(
+    throw UndefinedMemberD4rtException(
       "Undefined property or method '$name' on bridged instance of '${bridgedClass.name}'",
+      memberName: name,
     );
   }